refactor(target): make densetower a mlu-target pattern
